Create a `GetStructInfo` function to allow people to use the field info
This function is designed for people to use the info about the field
that this library generates internally for other uses that we thought
of.

Since structures can be tested, `GetStructInfo` takes a reflect.Type so
that when a field is a structure, it is possible to do
`GetStructInfo(fields[i].Type)` too.